likeCount, commentCount 도메인 모델의 필드로 접근해서 처리

+++ application/src/main/kotlin/com/jinuk/toy/applicaiton/comment/query/result/GetCommentPageResult.kt
@@ -7,7 +7,7 @@ data class GetCommentPageResult(
     val username: Username,
     val isViewerLike: Boolean,
     val parentCommentId: Long?,
-    val likeCount: Int,
+    val likeCount: Long,
     val content: String,
     val children: List<GetCommentPageResult>,
 )
